+++ pending-order-bot.php
@@ -3,7 +3,7 @@
  * Plugin Name: Pending Order Bot
  * Plugin URI:  https://github.com/badasswp/pending-order-bot
  * Description: Send reminders on WooCommerce pending orders.
- * Version:     1.0.1
+ * Version:     1.0.2
  * Author:      badasswp
  * Author URI:  https://github.com/badasswp
  * License:     GPL v2 or later
@@ -38,6 +38,9 @@ if ( ! file_exists( PBOT_AUTOLOAD ) ) {
 	return;
 }
 
-// Run Plugin.
+// Autoload classes.
 require_once PBOT_AUTOLOAD;
+
+// Run plugin.
+require_once __DIR__ . '/inc/Helpers/functions.php';
 ( \PendingOrderBot\Plugin::get_instance() )->run();