+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.2
+* Pass Twilio client via Dependency Injection.
+* Improve Unit tests & Test coverage.
+* Tested up to WP 6.7.2.
+
 ## 1.0.1
 * Restrict non-permited plugin files.
 * Update Twilio SDK package to 8.3.9.